gates, policies

+++ app/Http/Controllers/OwnerController.php
@@ -5,15 +5,26 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Owner;
 use App\Http\Requests\OwnerRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class OwnerController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $owners = Owner::all();
+        $user = auth()->user();
+
+        if (!$user) {
+            $owners = Owner::all();
+        } elseif ($user->privilege === 'admin' || $user->privilege === 'reader') {
+            $owners = Owner::all();
+        } else {
+            $owners = Owner::where('user_id', $user->id)->get();
+        }
+
         return view('owners.index', compact('owners'));
     }
 
@@ -30,7 +41,9 @@ class OwnerController extends Controller
      */
     public function store(OwnerRequest $request)
     {
-        Owner::create($request->all());
+        $data = $request->all();
+        $data['user_id'] = auth()->id();
+        Owner::create($data);
         return redirect()->route('owners.index');
     }
 
@@ -47,6 +60,7 @@ class OwnerController extends Controller
      */
     public function edit(Owner $owner)
     {
+        $this->authorize('update', $owner);
         return view('owners.edit', compact('owner'));
     }
 
@@ -55,12 +69,14 @@ class OwnerController extends Controller
      */
     public function update(OwnerRequest $request, Owner $owner)
     {
+        $this->authorize('update', $owner);
         $owner->update($request->all());
         return redirect()->route('owners.index');
     }
 
     public function destroy(Owner $owner)
     {
+        $this->authorize('delete', $owner);
         $owner->delete();
         return redirect()->route('owners.index');
     }
